+++ scripts/ci/test_profile_integration.sh
@@ -10,7 +10,7 @@ pip install -e ./tools
 [ -d privates ] && pip install -qqq privates/*.whl
 pip install $ALL_MODULES
 
-set -e
+set -ev
 
 output=$(az cloud list-profiles -otsv)