+++ src/azure-cli-core/azure/cli/core/tests/test_output.py
@@ -12,8 +12,9 @@ class TestCoreCLIOutput(unittest.TestCase):
         from azure.cli.core.mock import DummyCli
 
         output_producer = AzOutputProducer(DummyCli())
-        self.assertEqual(5, len(output_producer._FORMAT_DICT))  # five types: json, jsonc, table, tsv, yaml
+        self.assertEqual(6, len(output_producer._FORMAT_DICT))  # six types: json, jsonc, table, tsv, yaml, none
         self.assertIn('yaml', output_producer._FORMAT_DICT)
+        self.assertIn('none', output_producer._FORMAT_DICT)
 
 
 if __name__ == '__main__':