+++ src/command_modules/azure-cli-configure/azure/cli/command_modules/configure/_consts.py
@@ -9,7 +9,8 @@ OUTPUT_LIST = [
      'desc': 'Colored JSON formatted output that most closely matches API responses.'},
     {'name': 'table', 'desc': 'Human-readable output format.'},
     {'name': 'tsv', 'desc': 'Tab- and Newline-delimited. Great for GREP, AWK, etc.'},
-    {'name': 'yaml', 'desc': 'YAML formatted output. An alternative to JSON. Great for configuration files.'}
+    {'name': 'yaml', 'desc': 'YAML formatted output. An alternative to JSON. Great for configuration files.'},
+    {'name': 'none', 'desc': 'No output, except for errors and warnings.'}
 ]
 
 LOGIN_METHOD_LIST = [