+++ src/command_modules/azure-cli-storage/azure/cli/command_modules/storage/_params.py
@@ -5,7 +5,8 @@
 
 from azure.cli.core.profiles import ResourceType
 from azure.cli.core.commands.validators import get_default_location_from_resource_group
-from azure.cli.core.commands.parameters import (tags_type, file_type, get_location_type, get_enum_type)
+from azure.cli.core.commands.parameters import (tags_type, file_type, get_location_type, get_enum_type,
+                                                get_three_state_flag)
 
 from ._validators import (get_datetime_type, validate_metadata, get_permission_validator, get_permission_help_string,
                           resource_type_type, services_type, validate_entity, validate_select, validate_blob_type,
@@ -60,6 +61,9 @@ def load_arguments(self, _):  # pylint: disable=too-many-locals, too-many-statem
                                     action='store_true', validator=add_progress_callback)
     socket_timeout_type = CLIArgumentType(help='The socket timeout(secs), used by the service to regulate data flow.',
                                           type=int)
+    num_results_type = CLIArgumentType(
+        default=5000, help='Specifies the maximum number of results to return. Provide "*" to return all.',
+        validator=validate_storage_data_plane_list)
 
     sas_help = 'The permissions the SAS grants. Allowed values: {}. Do not use if a stored access policy is ' \
                'referenced with --id that specifies this value. Can be combined.'
@@ -205,11 +209,7 @@ def load_arguments(self, _):  # pylint: disable=too-many-locals, too-many-statem
 
     with self.argument_context('storage blob list') as c:
         c.argument('include', validator=validate_included_datasets)
-        c.argument('num_results', type=int,
-                   help='Specifies the maximum number of blobs to return. '
-                        'If this parameter is not provided, all blobs will be returned.',
-                   validator=validate_storage_data_plane_list)
-        c.ignore('marker')  # https://github.com/Azure/azure-cli/issues/3745
+        c.argument('num_results', arg_type=num_results_type)
 
     with self.argument_context('storage blob generate-sas') as c:
         from .completers import get_storage_acl_name_completion_list
@@ -247,6 +247,19 @@ def load_arguments(self, _):  # pylint: disable=too-many-locals, too-many-statem
         c.argument('enable', arg_type=get_enum_type(['true', 'false']), help='Enables/disables soft-delete.')
         c.argument('days_retained', type=int,
                    help='Number of days that soft-deleted blob will be retained. Must be in range [1,365].')
+
+    with self.argument_context('storage blob service-properties update', min_api='2018-03-28') as c:
+        c.argument('delete_retention', arg_type=get_three_state_flag(), arg_group='Soft Delete',
+                   help='Enables soft-delete.')
+        c.argument('delete_retention_period', type=int, arg_group='Soft Delete',
+                   help='Number of days that soft-deleted blob will be retained. Must be in range [1,365].')
+        c.argument('static_website', arg_group='Static Website', arg_type=get_three_state_flag(),
+                   help='Enables static-website.')
+        c.argument('index_document', help='Represents the name of the index document. This is commonly "index.html".',
+                   arg_group='Static Website')
+        c.argument('error_document_404_path', options_list=['--404-document'], arg_group='Static Website',
+                   help='Represents the path to the error document that should be shown when an error 404 is issued,'
+                        ' in other words, when a browser requests a page that does not exist.')
 
     with self.argument_context('storage blob upload') as c:
         from ._validators import page_blob_tier_validator
@@ -389,8 +402,7 @@ def load_arguments(self, _):  # pylint: disable=too-many-locals, too-many-statem
         c.ignore('blob_name', 'snapshot')
 
     with self.argument_context('storage container list') as c:
-        c.argument('num_results', type=int, help='Specifies the maximum number of containers to return.',
-                   validator=validate_storage_data_plane_list)
+        c.argument('num_results', arg_type=num_results_type)
 
     with self.argument_context('storage container set-permission') as c:
         c.ignore('signed_identifiers')
@@ -408,11 +420,8 @@ def load_arguments(self, _):  # pylint: disable=too-many-locals, too-many-statem
 
     with self.argument_context('storage container legal-hold') as c:
         c.argument('container_name', container_name_type)
-        c.argument('tags', nargs='+', help='Each tag should be 3 to 23 alphanumeric characters and is '
-                                           'normalized to lower case')
-
-    with self.argument_context('storage container list') as c:
-        c.ignore('marker')  # https://github.com/Azure/azure-cli/issues/3745
+        c.argument('tags', nargs='+',
+                   help='Each tag should be 3 to 23 alphanumeric characters and is normalized to lower case')
 
     with self.argument_context('storage container policy') as c:
         from .completers import get_storage_acl_name_completion_list
@@ -458,7 +467,7 @@ def load_arguments(self, _):  # pylint: disable=too-many-locals, too-many-statem
         c.argument('protocol', arg_type=get_enum_type(['http', 'https'], 'https'), help='Protocol to use.')
 
     with self.argument_context('storage share list') as c:
-        c.ignore('marker')  # https://github.com/Azure/azure-cli/issues/3745
+        c.argument('num_results', arg_type=num_results_type)
 
     with self.argument_context('storage share exists') as c:
         c.ignore('directory_name', 'file_name')
@@ -555,6 +564,7 @@ def load_arguments(self, _):  # pylint: disable=too-many-locals, too-many-statem
         from .completers import dir_path_completer
         c.argument('directory_name', options_list=('--path', '-p'), help='The directory path within the file share.',
                    completer=dir_path_completer)
+        c.argument('num_results', arg_type=num_results_type)
 
     with self.argument_context('storage file metadata show') as c:
         c.register_path_argument()