+++ src/command_modules/azure-cli-acs/HISTORY.rst
@@ -3,6 +3,11 @@
 Release History
 ===============
 
+2.3.14
+++++++
+* Minor fixes
+* Add support for new ACI regions
+
 2.3.13
 ++++++
 * Add Virtual Nodes Preview